Item has decimal value for amount

+++ app/src/main/java/com/example/supplytracker/Item.kt
@@ -4,4 +4,4 @@ package com.example.supplytracker
  * This item has a name and an amount that can be accessed or changed.
  * This class gets and sets the given name and the given amount for this item.
  */
-class Item(var name: String, var amount: Int)
+class Item(var name: String, var amount: Double)
